refactor: rename modals folder
