+++ src/components/Modals/ModalLogOut/styles.ts
