+++ src/components/Modals/ModalSocial/styles.ts
