+++ src/components/Modals/ModalSocialMediaMobile/styles.ts
