+++ src/components/Modals/ModalStakeAndWithdraw/styles.ts
