+++ src/components/Modals/ModalWalletConnect/index.tsx
@@ -83,7 +83,7 @@ const ModalWalletConnect = ({ setModalOpen }: IModalWalletConnect) => {
               </S.WrapperIconsBackGround>
             </Tippy>
 
-            <S.WrapperIconsBackGround
+            {/* <S.WrapperIconsBackGround
               type="button"
               onClick={() => {
                 setModalOpen(false)
@@ -94,7 +94,7 @@ const ModalWalletConnect = ({ setModalOpen }: IModalWalletConnect) => {
                 <img src="/assets/connectWalletIcon.svg" alt="" />
                 <span>WalletConnect</span>
               </S.WrapperIcons>
-            </S.WrapperIconsBackGround>
+            </S.WrapperIconsBackGround> */}
           </S.Content>
         </S.BackgroundBlack>
       </S.Container>
