+++ src/components/Modals/ModalWalletConnect/styles.ts
